Make some more 64-bit compatibility changes. Use correct types in Carbon methods and replace some deprecated methods or types.
git-svn-id: https://bibdesk.svn.sourceforge.net/svnroot/bibdesk/trunk/bibdesk@16067 0fceea05-a30d-0410-8a8b-80ef821fb0a1

+++ BDSKDocumentSearch.m
@@ -165,7 +165,7 @@ static NSOperationQueue *searchQueue = nil;
     search = SKSearchCreate(skIndex, (CFStringRef)searchString, kSKSearchOptionDefault);
     
     SKDocumentID documents[SEARCH_BUFFER_MAX] = { 0 };
-    CGFloat scores[SEARCH_BUFFER_MAX] = { 0.0 };
+    float scores[SEARCH_BUFFER_MAX] = { 0.0 };
     CFIndex i, foundCount;
     
     Boolean more, keepGoing;
